+++ src/__playground__/index.ts
@@ -1,8 +1,22 @@
 // 💡 index.ts
-import { ServiceCluster } from '@service_cluster';
 // Importing the Nanolith API we created in worker.ts
 import { worker } from './worker.js';
 
-const cluster = new ServiceCluster(worker);
+// Launch 6 services at the same time.
+const cluster = await worker.clusterize(6);
 
-const data = await cluster.launch(1);
+// Use the least busy service on the cluster.
+// This is the service that is currently running
+// the least amount of task calls.
+const service = cluster.use();
+
+// Call the task on the service as you normally would.
+const result = await service.call({
+    name: 'subtract',
+    params: [10, 5],
+});
+
+console.log(result);
+
+// Close all services on the cluster.
+await cluster.closeAll();