+++ src/__playground__/worker.ts
@@ -1,10 +1,17 @@
 // worker.ts 💼
 import { define } from 'nanolith';
 
-// Exporting the variable is not a requirement, but it is
-// necessary to somehow export the resolved value of the
-// function in order to have access to it later on.
 export const worker = await define({
+    __initializeService(threadId) {
+        console.log(`Initializing service on ${threadId}`);
+    },
+    __beforeTask({ name, inService }) {
+        console.log(`Running task ${name}.`);
+        console.log(`${inService ? 'Is' : 'Is not'} in a service.`);
+    },
+    __afterTask({ name, inService }) {
+        console.log(`Finished task ${name}`);
+    },
     add(x: number, y: number) {
         return x + y;
     },
@@ -12,9 +19,6 @@ export const worker = await define({
         await new Promise((resolve) => setTimeout(resolve, 5e3));
         return x + y;
     },
-    // Functions don't have to be directly defined within the
-    // object, they can be defined elsewhere outside, or even
-    // imported from a totally different module.
     subtract,
 });